+++ mobile/src/components/layout/index.ts
@@ -1,1 +1,3 @@
 export { default as HapticButton } from './HapticButton';
+export { default as ThemedView } from './ThemedView';
+export { default as TitledScreen } from './TitledScreen';
