+++ mobile/src/components/ui/index.ts
@@ -3,3 +3,5 @@ export { default as Pill } from './Pill';
 export { default as EmptyState } from './EmptyState';
 export { default as BottomSheet } from './BottomSheet';
 export { default as ThemedText } from './ThemedText';
+
+export { type BottomSheetModal } from '@gorhom/bottom-sheet';
